refactor(rpc): implement WalletRpc with maybe_async for compatibility
Add maybe_async support to WalletRpc trait enabling async/sync flexibility.
WalletRpc now implemented in floresta-node to ensure consistency.
