+++ crates/floresta-rpc/src/rpc_interfaces.rs
@@ -98,6 +98,7 @@ pub trait BlockchainRpc {
     ) -> Result<GetBlockHeaderRes, Self::Error>;
 }
 
+#[floresta_proc_macro::maybe_async]
 pub trait WalletRpc {
     type Error: Display + Debug;
